Refactor test for sourceRecorder
