+++ src/features/sourceRecorder/sourcereel/__tests__/SourcereelActions.ts
@@ -3,8 +3,7 @@ import {
   WorkspaceLocation,
   WorkspaceLocations
 } from '../../../../commons/workspace/WorkspaceTypes';
-import { saveSourcecastData } from '../../SourceRecorderActions';
-import { CodeDelta, Input, PlaybackData, SAVE_SOURCECAST_DATA } from '../../SourceRecorderTypes';
+import { CodeDelta, Input, PlaybackData } from '../../SourceRecorderTypes';
 import {
   recordInit,
   recordInput,
@@ -69,35 +68,6 @@ test('recordInput generates correct action object', () => {
     type: RECORD_INPUT,
     payload: {
       input,
-      workspaceLocation: sourcereelWorkspace
-    }
-  });
-});
-
-test('saveSourcecastData generates correct action object', () => {
-  const fakeUrl = 'someFakeAudioUrl.com';
-  const noOp = () => fakeUrl;
-  window.URL.createObjectURL = noOp;
-  const title = 'Test Title';
-  const description = 'Test Description';
-  const audio = new Blob();
-  const playbackData: PlaybackData = {
-    init: {
-      editorValue: 'Editor Init Value',
-      chapter: 1,
-      externalLibrary: ExternalLibraryNames.NONE
-    },
-    inputs: []
-  };
-  const action = saveSourcecastData(title, description, audio, playbackData, sourcereelWorkspace);
-  expect(action).toEqual({
-    type: SAVE_SOURCECAST_DATA,
-    payload: {
-      title,
-      description,
-      audio,
-      audioUrl: fakeUrl,
-      playbackData,
       workspaceLocation: sourcereelWorkspace
     }
   });
